refactor: move otp types to wirex types
